JiraAgent/pyproject.toml
[project]
name = "jira-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.5",
    "langchain-core>=1.0.4",
    "langchain-google-genai>=3.0.3",
    "langchain-mcp-adapters>=0.1.13",
    "langgraph>=1.0.3",
    "matplotlib>=3.10.7",
    "mcp>=1.21.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
]
